Title: 정수 세 개 입력받아 출력, Time: 42ms, Memory: 16MB, Status: Wrong Answer - CodeTree

+++ 250123/정수 세 개 입력받아 출력/take-three-integers-and-output.py
@@ -3,7 +3,7 @@ arr = n.split()
 
 a = int(arr[0])
 b = int(arr[1])
-c = float(input())
+c = int(input())
 
 print(a, b)
 print(c)
